+++ zerver/lib/markdown/__init__.py
@@ -48,6 +48,7 @@ from zerver.lib.markdown import fenced_code
 from zerver.lib.markdown.fenced_code import FENCE_RE
 from zerver.lib.mention import (
     BEFORE_MENTION_ALLOWED_REGEX,
+    ChannelTopicInfo,
     FullNameInfo,
     MentionBackend,
     MentionData,
@@ -130,6 +131,7 @@ class DbData:
     active_realm_emoji: dict[str, EmojiInfo]
     sent_by_bot: bool
     stream_names: dict[str, int]
+    topic_info: dict[ChannelTopicInfo, int | None]
     translate_emoticons: bool
     user_upload_previews: dict[str, MarkdownImageMetadata]
 
@@ -2049,6 +2051,13 @@ class StreamPattern(StreamTopicMessageProcessor):
 
 
 class StreamTopicPattern(StreamTopicMessageProcessor):
+    def get_with_operand(self, channel_topic: ChannelTopicInfo) -> int | None:
+        db_data: DbData | None = self.zmd.zulip_db_data
+        if db_data is None:
+            return None
+        with_operand = db_data.topic_info.get(channel_topic)
+        return with_operand
+
     @override
     def handleMatch(  # type: ignore[override] # https://github.com/python/mypy/issues/10197
         self, m: Match[str], data: str
@@ -2064,7 +2073,13 @@ class StreamTopicPattern(StreamTopicMessageProcessor):
         el.set("data-stream-id", str(stream_id))
         stream_url = encode_stream(stream_id, stream_name)
         topic_url = hash_util_encode(topic_name)
-        link = f"/#narrow/channel/{stream_url}/topic/{topic_url}"
+        channel_topic_object = ChannelTopicInfo(stream_name, topic_name)
+        with_operand = self.get_with_operand(channel_topic_object)
+        if with_operand is not None:
+            link = f"/#narrow/channel/{stream_url}/topic/{topic_url}/with/{with_operand}"
+        else:
+            link = f"/#narrow/channel/{stream_url}/topic/{topic_url}"
+
         el.set("href", link)
 
         if topic_name == "":
@@ -2128,6 +2143,15 @@ def possible_linked_stream_names(content: str) -> set[str]:
         # `STREAM_TOPIC_MESSAGE_LINK_REGEX` here, but the way channel
         # names are separated, STREAM_TOPIC_LINK_REGEX will have
         # already found them.
+    }
+
+
+def possible_linked_topics(content: str) -> set[ChannelTopicInfo]:
+    # Here, we do not consider STREAM_TOPIC_MESSAGE_LINK_REGEX, since
+    # our callers only want to process links without a message ID.
+    return {
+        ChannelTopicInfo(match.group("stream_name"), match.group("topic_name"))
+        for match in re.finditer(STREAM_TOPIC_LINK_REGEX, content, re.VERBOSE)
     }
 
 
@@ -2740,15 +2764,21 @@ def do_convert(
         # the fetches are somewhat expensive and these types of syntax
         # are uncommon enough that it's a useful optimization.
 
+        message_sender = None
+        if message is not None:
+            message_sender = message.sender
+
         if mention_data is None:
             mention_backend = MentionBackend(message_realm.id)
-            message_sender = None
-            if message is not None:
-                message_sender = message.sender
             mention_data = MentionData(mention_backend, content, message_sender)
 
         stream_names = possible_linked_stream_names(content)
-        stream_name_info = mention_data.get_stream_name_map(stream_names)
+        stream_name_info = mention_data.get_stream_name_map(
+            stream_names, acting_user=message_sender
+        )
+
+        linked_stream_topic_data = possible_linked_topics(content)
+        topic_info = mention_data.get_topic_info_map(linked_stream_topic_data)
 
         if content_has_emoji_syntax(content):
             active_realm_emoji = get_name_keyed_dict_for_active_realm_emoji(message_realm.id)
@@ -2763,6 +2793,7 @@ def do_convert(
             realm_url=message_realm.url,
             sent_by_bot=sent_by_bot,
             stream_names=stream_name_info,
+            topic_info=topic_info,
             translate_emoticons=translate_emoticons,
             user_upload_previews=user_upload_previews,
         )